+++ PROYECTO_ENCOMIENDA/nbproject/genfiles.properties
@@ -1,8 +1,8 @@
-build.xml.data.CRC32=8732ab98
+build.xml.data.CRC32=b86db72a
 build.xml.script.CRC32=65fda716
-build.xml.stylesheet.CRC32=f85dc8f2@1.109.0.48
+build.xml.stylesheet.CRC32=f85dc8f2@1.111.0.48
 # This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
 # Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=8732ab98
-nbproject/build-impl.xml.script.CRC32=85c0e8ec
-nbproject/build-impl.xml.stylesheet.CRC32=12e0a6c2@1.109.0.48
+nbproject/build-impl.xml.data.CRC32=b86db72a
+nbproject/build-impl.xml.script.CRC32=9b545093
+nbproject/build-impl.xml.stylesheet.CRC32=12e0a6c2@1.111.0.48
