+++ PROYECTO_ENCOMIENDA/nbproject/private/private.properties
@@ -5,4 +5,4 @@ do.jlink=false
 javac.debug=true
 javadoc.preview=true
 jlink.strip=false
-user.properties.file=C:\\Users\\Rodrigo Haro\\AppData\\Roaming\\NetBeans\\20\\build.properties
+user.properties.file=C:\\Users\\paulc\\AppData\\Roaming\\NetBeans\\22\\build.properties
