+++ PROYECTO_ENCOMIENDA/nbproject/private/private.xml
@@ -3,13 +3,12 @@
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/C:/Users/migue/Documents/GitHub/ProyectoPaquete/ProyectoPaquetes/src/GUICONDUCTOR/JFPaquetesCliente.java</file>
-            <file>file:/C:/Users/migue/Documents/GitHub/ProyectoPaquete/ProyectoPaquetes/src/GUI/JFVehiculo.java</file>
-            <file>file:/C:/Users/migue/Documents/GitHub/ProyectoPaquete/ProyectoPaquetes/src/GUI/JFMenu.java</file>
-            <file>file:/C:/Users/migue/Documents/GitHub/ProyectoPaquete/ProyectoPaquetes/src/GUI/JFRecuperar.java</file>
-            <file>file:/C:/Users/migue/Documents/GitHub/ProyectoPaquete/ProyectoPaquetes/src/GUICONDUCTOR/JFMenuConductor.java</file>
-            <file>file:/C:/Users/migue/Documents/GitHub/ProyectoPaquete/ProyectoPaquetes/src/proyecto_paquetes/JFIngresar.java</file>
-            <file>file:/C:/Users/migue/Documents/GitHub/ProyectoPaquete/ProyectoPaquetes/src/GUICONDUCTOR/JFPaquetesConductor.java</file>
+            <file>file:/C:/Users/paulc/OneDrive/Documentos/GitHub/ProyectoPaquetes/src/mod_transporte/AsignacionConductor.java</file>
+            <file>file:/C:/Users/paulc/OneDrive/Documentos/GitHub/ProyectoPaquetes/src/GUI/JFVehiculo.java</file>
+            <file>file:/C:/Users/paulc/OneDrive/Documentos/GitHub/ProyectoPaquetes/src/mod_transporte/AsignacionPaquete.java</file>
+            <file>file:/C:/Users/paulc/OneDrive/Documentos/GitHub/ProyectoPaquetes/src/mod_transporte/FlotaVehiculo.java</file>
+            <file>file:/C:/Users/paulc/OneDrive/Documentos/GitHub/ProyectoPaquetes/src/mod_administracion/Recepcionista.java</file>
+            <file>file:/C:/Users/paulc/OneDrive/Documentos/GitHub/ProyectoPaquetes/src/proyecto_paquetes/Proyecto_paquetes.java</file>
         </group>
     </open-files>
 </project-private>
