+++ PROYECTO_ENCOMIENDA/nbproject/project.properties
@@ -155,7 +155,7 @@ main.class=proyecto_paquetes.Proyecto_paquetes
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
-platform.active=JDK_22
+platform.active=default_platform
 run.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
