+++ PROYECTO_ENCOMIENDA/src/GUICONDUCTOR/JFIncidenteConductor.java
@@ -24,7 +24,6 @@ import mod_paquetes.Inventario;
 import mod_paquetes.Paquete;
 import mod_paquetes.Pendiente;
 import mod_paquetes.Seguimiento;
-import mod_transporte.Asignacion;
 import validaciones.*;
 
 /**
@@ -390,7 +389,7 @@ public class JFIncidenteConductor extends javax.swing.JFrame {
                 "Registro Exitoso",
                 JOptionPane.INFORMATION_MESSAGE);
             Inventario.obtenerInstancia().guardarInventario();
-            Asignacion.obtenerInstancia().guardarRelacionPaquetes();
+            conductor.obtenerAsignacionPaquete().guardarRelacionPaquetes();
             DefaultTableModel modeloTabla = (DefaultTableModel) jTablaPaquete.getModel();
             modeloTabla.setRowCount(0);
         } else {
@@ -509,7 +508,7 @@ public class JFIncidenteConductor extends javax.swing.JFrame {
                 "Resolución Exitosa",
                 JOptionPane.INFORMATION_MESSAGE);
             Inventario.obtenerInstancia().guardarInventario();
-            Asignacion.obtenerInstancia().guardarRelacionPaquetes();
+            conductor.obtenerAsignacionPaquete().guardarRelacionPaquetes();
             DefaultTableModel modeloTabla = (DefaultTableModel) jTIncidente.getModel();
             modeloTabla.setRowCount(0);
         } else {
