+++ PROYECTO_ENCOMIENDA/src/mod_administracion/Conductor.java
@@ -3,13 +3,14 @@ package mod_administracion;
 import mod_paquetes.EnCurso;
 import mod_paquetes.EstadoDelPaquete;
 import mod_paquetes.Paquete;
-import mod_transporte.Asignacion;
 
 import java.util.ArrayList;
 import mod_incidentes.PaqueteNoTieneIncidente;
 import mod_incidentes.PaqueteYaTieneIncidente;
 import mod_paquetes.Entregado;
 import mod_paquetes.Inventario;
+import mod_transporte.AsignacionConductor;
+import mod_transporte.Vehiculo;
 
 /**
  * Representa un conductor que puede reportar incidentes y consultar paquetes
@@ -60,7 +61,11 @@ public class Conductor extends Usuario {
      */
 
     public ArrayList<Paquete> consultarPaquetesAsignados() {
-        return Asignacion.obtenerInstancia().obtenerPaquetesDeConductor(this);
+        Vehiculo vehiculo = asignacionConductor.obtenerVehiculoDeConductor(this);
+        if(vehiculo == null){
+            return null;
+        }
+        return asignacionPaquete.obtenerPaquetesVehiculo(vehiculo);
     }
 
     public void entregarPaquete(String codigo) {
@@ -68,6 +73,6 @@ public class Conductor extends Usuario {
         paqueteCambio.cambiarEstado(new Entregado(paqueteCambio));
 
         Inventario.obtenerInstancia().guardarInventario();
-        Asignacion.obtenerInstancia().guardarRelacionPaquetes();
+        asignacionPaquete.guardarRelacionPaquetes();
     }
 }
