+++ PROYECTO_ENCOMIENDA/src/proyecto_paquetes/Proyecto_paquetes.java
@@ -6,7 +6,7 @@ package proyecto_paquetes;
 
 import mod_facturacion.Cotizacion;
 import mod_paquetes.Inventario;
-import mod_transporte.Asignacion;
+import mod_transporte.FlotaVehiculo;
 
 /**
  *
@@ -18,10 +18,6 @@ public class Proyecto_paquetes {
      * @param args the command line arguments
      */
     public static void main(String[] args) {
-        Asignacion.obtenerInstancia().cargarVehiculos();
-        Asignacion.obtenerInstancia().cargarConductores();
-        Asignacion.obtenerInstancia().cargarRelacionConductores();
-        Asignacion.obtenerInstancia().cargarRelacionPaquetes();
         Inventario.obtenerInstancia().cargarInventario();
         Cotizacion.obtenerInstancia().cargarCotizacion();
         JFIngresar loginForm = new JFIngresar();
